refactor(core): Move root scripts to src/ and fix all imports
