--- main.py
@@ -1,534 +1,0 @@
-"""
-main.py
-
-This module serves as the entry point for the ChattyCommander application. It coordinates the
-loading of machine learning models, manages state transitions based on voice commands, and
-handles the execution of commands.
-
-Usage:
-    Run the script from the command line to start the voice-activated command processing system.
-    Ensure that all dependencies are installed and models are correctly placed in their respective directories.
-
-Example:
-    python main.py
-"""
-
-import argparse
-
-# Ensure src/ is on sys.path so root execution finds package modules without PYTHONPATH
-import os as _os
-import signal
-import sys
-import sys as _sys
-import threading
-
-_src_path = _os.path.join(_os.path.dirname(_os.path.abspath(__file__)), "src")
-if _src_path not in _sys.path:
-    _sys.path.insert(0, _src_path)
-
-# Support both package and repo-root execution without PYTHONPATH tweaks.
-try:
-    # Preferred: installed package
-    from chatty_commander.app.command_executor import CommandExecutor  # type: ignore
-    from chatty_commander.app.model_manager import ModelManager  # type: ignore
-    from chatty_commander.app.state_manager import StateManager  # type: ignore
-    from chatty_commander.app.orchestrator import (  # type: ignore
-        ModeOrchestrator,
-        OrchestratorFlags,
-    )
-    from chatty_commander.config import Config  # type: ignore
-    from chatty_commander.utils.logger import setup_logger  # type: ignore
-except Exception:
-    # Repo-root fallback: use local shim modules that re-export src implementations
-    from command_executor import CommandExecutor  # shim file at repo root
-    from config import Config  # shim file at repo root
-    from model_manager import ModelManager  # shim file at repo root
-    from state_manager import StateManager  # shim file at repo root
-    from utils.logger import setup_logger  # local utils
-    # Orchestrator shipped under src/chatty_commander/app
-    from chatty_commander.app.orchestrator import (  # type: ignore
-        ModeOrchestrator,
-        OrchestratorFlags,
-    )
-
-try:
-    from default_config import generate_default_config_if_needed
-except ImportError:
-
-    def generate_default_config_if_needed() -> bool:
-        return False
-
-
-def run_cli_mode(config, model_manager, state_manager, command_executor, logger):
-    """Run the traditional CLI voice command mode with graceful shutdown."""
-    logger.info("Starting CLI voice command mode")
-
-    # Load models based on the initial idle state
-    model_manager.reload_models(state_manager.current_state)
-
-    shutdown_flag = {"stop": False}
-
-    def handle_signal(signum, frame):
-        logger.info(f"Received signal {signum}, initiating graceful shutdown...")
-        shutdown_flag["stop"] = True
-
-    # Register signal handlers for graceful shutdown
-    signal.signal(signal.SIGINT, handle_signal)
-    signal.signal(signal.SIGTERM, handle_signal)
-
-    try:
-        while not shutdown_flag["stop"]:
-            # Listen for voice input
-            command = model_manager.listen_for_commands()
-            if not command:
-                continue
-
-            logger.info(f"Command detected: {command}")
-
-            # Update system state based on command
-            new_state = state_manager.update_state(command)
-            if new_state:
-                logger.info(f"Transitioning to new state: {new_state}")
-                model_manager.reload_models(new_state)
-
-            # Execute the detected command if it's actionable
-            if command in config.model_actions:
-                command_executor.execute_command(command)
-
-    except KeyboardInterrupt:
-        logger.info("KeyboardInterrupt received; shutting down")
-    finally:
-        # Perform any resource cleanup if needed
-        try:
-            if hasattr(model_manager, "shutdown"):
-                model_manager.shutdown()
-            if hasattr(state_manager, "shutdown"):
-                state_manager.shutdown()
-        except Exception as e:
-            logger.error(f"Error during shutdown: {e}")
-        logger.info("ChattyCommander CLI shutdown complete")
-        sys.exit(0)
-
-
-def run_web_mode(
-    config, model_manager, state_manager, command_executor, logger, no_auth=False, port=8100
-):
-    """Run the web UI mode with FastAPI server and graceful shutdown."""
-    try:
-        from web_mode import create_web_server
-    except ImportError:
-        logger.error(
-            "Web mode dependencies not available. Install with: uv add fastapi uvicorn websockets"
-        )
-        sys.exit(1)
-
-    logger.info(f"Starting web mode (auth={'disabled' if no_auth else 'enabled'}) on port {port}")
-
-    # Create web server instance
-    web_server = create_web_server(
-        config_manager=config,
-        state_manager=state_manager,
-        model_manager=model_manager,
-        command_executor=command_executor,
-        no_auth=no_auth,
-    )
-
-    # Setup callbacks for voice command integration
-    def on_command_detected(command):
-        web_server.on_command_detected(command)
-
-    def on_state_change(old_state, new_state):
-        web_server._on_state_change(old_state, new_state)
-
-    # Register callbacks
-    if hasattr(model_manager, 'add_command_callback'):
-        model_manager.add_command_callback(on_command_detected)
-    if hasattr(state_manager, 'add_state_change_callback'):
-        state_manager.add_state_change_callback(on_state_change)
-
-    stop_event = threading.Event()
-
-    def handle_signal(signum, frame):
-        logger.info(f"Received signal {signum}, stopping web server...")
-        stop_event.set()
-        try:
-            stopper = getattr(web_server, "stop", None)
-            if callable(stopper):
-                stopper()
-        except Exception as e:
-            logger.error(f"Error stopping web server: {e}")
-
-    signal.signal(signal.SIGINT, handle_signal)
-    signal.signal(signal.SIGTERM, handle_signal)
-
-    # Start the server
-    try:
-        web_server.run(host="0.0.0.0", port=port, log_level="info")
-    finally:
-        try:
-            if hasattr(model_manager, "shutdown"):
-                model_manager.shutdown()
-            if hasattr(state_manager, "shutdown"):
-                state_manager.shutdown()
-        except Exception as e:
-            logger.error(f"Error during web mode shutdown: {e}")
-        logger.info("Web mode shutdown complete")
-
-
-def run_gui_mode(
-    config,
-    model_manager,
-    state_manager,
-    command_executor,
-    logger,
-    display_override: str | None = None,
-    no_gui: bool = False,
-):
-    """Run the GUI mode with graceful handling in headless environments.
-
-    Returns:
-        int: 0 if skipped or exited cleanly; non-zero if GUI could not start due to missing deps.
-    """
-    import os
-
-    if no_gui:
-        logger.info("--no-gui specified; skipping GUI launch")
-        return 0
-
-    # Apply DISPLAY override if provided (POSIX only)
-    if display_override and os.name != "nt":
-        os.environ["DISPLAY"] = display_override
-
-    # Graceful handling when DISPLAY is not present (headless/CI on POSIX)
-    if os.name != "nt" and not os.environ.get("DISPLAY"):
-        logger.warning(
-            "No DISPLAY environment variable set. Skipping GUI mode in headless environment."
-        )
-        return 0
-    # Prefer new tray popup GUI (pystray + pywebview) with fallback to legacy tkinter GUI
-    try:
-        try:
-            # Installed package path
-            from chatty_commander.gui.tray_popup import run_tray_popup  # type: ignore
-        except Exception:
-            # Repo-root execution fallback
-            from src.chatty_commander.gui.tray_popup import run_tray_popup  # type: ignore
-
-        logger.info("Starting GUI tray popup mode")
-        rc = run_tray_popup(config, logger)
-        return 0 if rc is None else int(rc)
-    except Exception as e:
-        logger.warning(f"Tray popup GUI unavailable ({e}); falling back to legacy tkinter GUI")
-        try:
-            from gui import main as gui_main
-
-            logger.info("Starting legacy tkinter GUI mode")
-            rc = gui_main()
-            return 0 if rc is None else int(rc)
-        except Exception:
-            logger.error("GUI dependencies not available. Install with: uv add tkinter")
-            return 2
-
-
-def create_parser():
-    """Create and configure the argument parser."""
-    parser = argparse.ArgumentParser(
-        description="ChattyCommander - Advanced voice-activated command processing system.\n"
-        "This application allows users to control their computer using voice commands, "
-        "with support for multiple modes including CLI, web UI, GUI, and configuration wizard.\n"
-        "It integrates machine learning models for command detection and state management.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  %(prog)s                    # Start in CLI voice command mode
-  %(prog)s --shell            # Start interactive text shell mode
-  %(prog)s --web              # Start web UI server on default port 8100
-  %(prog)s --web --port 8080 --no-auth  # Start web server on port 8080 without auth
-  %(prog)s --gui              # Launch graphical user interface
-  %(prog)s --config           # Run interactive configuration wizard
-  %(prog)s --log-level DEBUG  # Start with debug logging
-
-Available modes:
-- CLI: Voice-activated command processing
-- Shell: Interactive text-based command input
-- Web: Browser-based interface with real-time updates
-- GUI: Desktop application interface
-- Config: Setup and configuration tool
-
-For detailed documentation and source code, visit: https://github.com/your-repo/chatty-commander
-        """,
-    )
-
-    mode_group = parser.add_mutually_exclusive_group()
-    mode_group.add_argument(
-        "--web",
-        action="store_true",
-        help="Start the web UI server using FastAPI backend. Requires FastAPI and Uvicorn. "
-        "Serves a React-based frontend if built.",
-    )
-    mode_group.add_argument(
-        "--gui",
-        action="store_true",
-        help="Start the graphical user interface. Requires Tkinter and a DISPLAY environment.",
-    )
-    mode_group.add_argument(
-        "--config",
-        action="store_true",
-        help="Launch the interactive configuration wizard to set up models, commands, and settings.",
-    )
-    mode_group.add_argument(
-        "--shell",
-        action="store_true",
-        help="Start interactive shell mode for text-based command input and execution.",
-    )
-
-    # Orchestrator flags (opt-in; does not change default behavior)
-    parser.add_argument(
-        "--orchestrate",
-        action="store_true",
-        help="Use the mode orchestrator to unify adapters (text, web, gui, wakeword, cv, discord bridge).",
-    )
-    parser.add_argument(
-        "--enable-text",
-        action="store_true",
-        help="Enable text adapter in orchestrator.",
-    )
-    parser.add_argument(
-        "--enable-openwakeword",
-        action="store_true",
-        help="Enable OpenWakeWord adapter in orchestrator.",
-    )
-    parser.add_argument(
-        "--enable-computer-vision",
-        action="store_true",
-        help="Enable Computer Vision adapter in orchestrator.",
-    )
-    parser.add_argument(
-        "--enable-discord-bridge",
-        action="store_true",
-        help="Enable Discord/Slack bridge adapter (requires advisors + bridge token).",
-    )
-
-    parser.add_argument(
-        "--no-auth",
-        action="store_true",
-        help="Disable authentication for web mode (INSECURE - use only for local development).",
-    )
-
-    parser.add_argument(
-        "--port",
-        type=int,
-        default=8100,
-        help="Specify the port for the web server (default: 8100). Only used in web mode.",
-    )
-
-    parser.add_argument(
-        "--log-level",
-        choices=["DEBUG", "INFO", "WARNING", "ERROR"],
-        default="INFO",
-        help="Set the logging level for the application (default: INFO).",
-    )
-
-    # Headless-friendly switches
-    parser.add_argument(
-        "--no-gui",
-        action="store_true",
-        help="Avoid launching the GUI even if --gui is provided; useful in CI/headless.",
-    )
-    parser.add_argument(
-        "--display", type=str, default=None, help="Override DISPLAY value for GUI mode (e.g., :0)."
-    )
-
-    # Advisors convenience flag
-    parser.add_argument(
-        "--advisors",
-        action="store_true",
-        help="Enable advisors feature at runtime (overrides config.advisors.enabled).",
-    )
-
-    return parser
-
-
-def run_interactive_shell(config, model_manager, state_manager, command_executor, logger):
-    """Run interactive text-based shell mode with tab completion."""
-    import readline
-
-    logger.info("Starting interactive shell mode")
-    print("ChattyCommander Interactive Shell")
-    print("Type 'help' for commands, 'exit' to quit")
-
-    commands = ['help', 'exit', 'state', 'models', 'execute']
-    model_actions = list(config.model_actions.keys())
-
-    def completer(text, state):
-        options = [cmd for cmd in commands if cmd.startswith(text)]
-        if text.startswith('execute '):
-            subtext = text[8:]
-            suboptions = [f'execute {act}' for act in model_actions if act.startswith(subtext)]
-            try:
-                return suboptions[state]
-            except IndexError:
-                return None
-        try:
-            return options[state]
-        except IndexError:
-            return None
-
-    readline.set_completer(completer)
-    readline.parse_and_bind('tab: complete')
-
-    while True:
-        try:
-            input_str = input("> ").strip()
-            if not input_str:
-                continue
-            if input_str.lower() == 'exit':
-                break
-            if input_str.lower() == 'help':
-                print("Available commands: help, exit, state, models, execute <command>")
-                continue
-            if input_str.lower() == 'state':
-                print(f"Current state: {state_manager.current_state}")
-                continue
-            if input_str.lower() == 'models':
-                print("Loaded models: " + ", ".join(model_manager.get_models()))
-                continue
-            if input_str.startswith('execute '):
-                command = input_str[8:].strip()
-                if command in config.model_actions:
-                    command_executor.execute_command(command)
-                    print(f"Executed: {command}")
-                else:
-                    print(f"Unknown command: {command}")
-                continue
-
-            # Treat as voice command simulation
-            new_state = state_manager.update_state(input_str)
-            if new_state:
-                logger.info(f"Transitioning to new state: {new_state}")
-                model_manager.reload_models(new_state)
-            if input_str in config.model_actions:
-                command_executor.execute_command(input_str)
-        except EOFError:
-            break
-    logger.info("Exiting interactive shell")
-
-
-def run_orchestrator_mode(config, model_manager, state_manager, command_executor, logger, args):
-    """Run orchestrator-driven mode; adapters route to the same command sink."""
-    flags = OrchestratorFlags(
-        enable_text=bool(getattr(args, "enable_text", False)),
-        enable_gui=bool(getattr(args, "gui", False)),
-        enable_web=bool(getattr(args, "web", False)),
-        enable_openwakeword=bool(getattr(args, "enable_openwakeword", False)),
-        enable_computer_vision=bool(getattr(args, "enable_computer_vision", False)),
-        enable_discord_bridge=bool(getattr(args, "enable_discord_bridge", False)),
-    )
-    orchestrator = ModeOrchestrator(
-        config=config,
-        command_sink=command_executor,
-        advisor_sink=None,
-        flags=flags,
-    )
-    selected = orchestrator.start()
-    logger.info(f"Orchestrator started adapters: {selected}")
-    # For now, block on CLI loop to keep process alive if no web/gui
-    if not args.web and not args.gui:
-        try:
-            while True:
-                signal.pause()
-        except KeyboardInterrupt:
-            pass
-    orchestrator.stop()
-    return 0
-
-
-def main():
-    parser = create_parser()
-    # Parse as the very first action and immediately return argparse exit code for help/usage
-    # We must allow --help to exit(0) without doing any setup, to satisfy tests.
-    try:
-        args, _unknown = parser.parse_known_args()
-    except SystemExit as e:
-        # Propagate argparse's exit code (0 on --help)
-        return int(getattr(e, "code", 0) or 0)
-
-    # If help was requested, argparse would have exited above with code 0.
-    # Continue with validation for actual runs only.
-    # Argument validation (only enforce when options are provided)
-    if getattr(args, "web", False) and getattr(args, "port", 8100) < 1024:
-        parser.error("Port must be 1024 or higher for non-root users")
-    if getattr(args, "no_auth", False) and not getattr(args, "web", False):
-        parser.error("--no-auth only applicable in web mode")
-
-    # If user only asked for help (--help), we would have already returned.
-    # If no args other than program name, print intro and exit 0 per tests expecting non-crash and intro visibility.
-    if len(sys.argv) == 1:
-        print("ChattyCommander - Voice Command System")
-        print("Use --help for available options")
-        # Align with tests expecting SystemExit on main invocation path.
-        raise SystemExit(0)
-
-    # Ensure logger is created with the expected name for tests
-    logger = setup_logger('main', 'logs/chattycommander.log')
-    logger.info("Starting ChattyCommander application")
-
-    # Generate default configuration if needed
-    if generate_default_config_if_needed():
-        logger.info("Default configuration generated")
-
-    # Load configuration settings
-    config = Config()
-    # Apply runtime advisors enable if requested
-    if getattr(args, "advisors", False):
-        try:
-            if not hasattr(config, "advisors"):
-                config.advisors = {}
-            config.advisors["enabled"] = True
-        except Exception:
-            pass
-    model_manager = ModelManager(config)
-    state_manager = StateManager()
-    command_executor = CommandExecutor(config, model_manager, state_manager)
-
-    # Route to appropriate mode
-    if getattr(args, "config", False):
-        from config_cli import ConfigCLI
-
-        config_cli = ConfigCLI()
-        config_cli.run_wizard()
-        return 0
-    elif getattr(args, "web", False):
-        run_web_mode(
-            config, model_manager, state_manager, command_executor, logger, args.no_auth, args.port
-        )
-        return 0
-    elif args.gui:
-        # Respect return codes from GUI runner (0=headless skipped, 2=deps missing)
-        rc = run_gui_mode(
-            config,
-            model_manager,
-            state_manager,
-            command_executor,
-            logger,
-            display_override=args.display,
-            no_gui=args.no_gui,
-        )
-        if isinstance(rc, int) and rc != 0:
-            # Non-zero means GUI could not start; exit without stack trace
-            return rc
-        return 0
-    elif args.shell:
-        run_interactive_shell(config, model_manager, state_manager, command_executor, logger)
-        return 0
-    elif getattr(args, "orchestrate", False):
-        return run_orchestrator_mode(
-            config, model_manager, state_manager, command_executor, logger, args
-        )
-    else:
-        run_cli_mode(config, model_manager, state_manager, command_executor, logger)
-        return 0
-
-
-if __name__ == "__main__":
-    sys.exit(main())
