+++ src/chatty_commander/default_config.py
