+++ src/chatty_commander/model_manager.py
