+++ src/chatty_commander/state_manager.py
