+++ tests/test_config.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from config import Config
+from src.chatty_commander.config import Config
 
 
 @pytest.fixture
