+++ tests/test_helpers.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from helpers import ensure_directory_exists, format_command_output, parse_model_keybindings
+from src.chatty_commander.helpers import ensure_directory_exists, format_command_output, parse_model_keybindings
 
 
 class TestHelpers(unittest.TestCase):
