+++ tests/test_integration_voice.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from config import Config
+from src.chatty_commander.config import Config
 from model_manager import ModelManager
 from state_manager import StateManager
