+++ tests/test_model_manager.py
@@ -1,6 +1,6 @@
 from unittest.mock import MagicMock, patch
 
-from config import Config
+from src.chatty_commander.config import Config
 from model_manager import ModelManager
