+++ tests/test_performance_benchmarks.py
@@ -14,7 +14,7 @@ import psutil
 import pytest
 from fastapi.testclient import TestClient
 
-from command_executor import CommandExecutor
+from src.chatty_commander.command_executor import CommandExecutor
 from config import Config
 from model_manager import ModelManager
 from state_manager import StateManager
