+++ tests/test_web_mode_unit.py
@@ -14,7 +14,7 @@ from unittest.mock import AsyncMock, Mock, patch, MagicMock
 import pytest
 from fastapi.testclient import TestClient
 
-from command_executor import CommandExecutor
+from src.chatty_commander.command_executor import CommandExecutor
 from config import Config
 from model_manager import ModelManager
 from state_manager import StateManager
